# docs/en/api/persona-v1.openapi.yaml
openapi: 3.1.0
info:
  title: SophIA Persona Engine API
  version: 1.2.0
  description: >-
    Loopback-only contract used by the SophIA desktop runtime. Version 1 exposes
    service readiness, immutable persona snapshots, turn and action context,
    and selective memory management. It does not execute tools, generate
    responses, or synthesize speech.
servers:
  - url: http://127.0.0.1:8765
security: []
paths:
  /v1/health:
    get:
      operationId: getPersonaHealth
      summary: Check Persona Engine readiness
      responses:
        "200":
          description: The service and required local components are ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: The service is alive but a required component is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/personas/{character_id}/snapshot:
    get:
      operationId: getPersonaSnapshot
      summary: Build the current immutable persona snapshot
      parameters:
        - name: character_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Snapshot built from the current character, state, and active user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonaSnapshot"
        "404":
          description: No character exists with the requested identifier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: The stored character cannot satisfy the persona contract.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Persona persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/personas/{character_id}/turn-context:
    post:
      operationId: buildPersonaTurnContext
      summary: Build a persona snapshot with relevant continuous memory
      parameters:
        - name: character_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TurnContextRequest"
      responses:
        "200":
          description: Persona snapshot augmented with bounded relevant memory.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonaSnapshot"
        "404":
          description: No character exists with the requested identifier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: The request or stored character is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Persona or memory persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/personas/{character_id}/action-context:
    post:
      operationId: buildPersonaActionContext
      summary: Build persona context around an authoritative action result
      parameters:
        - name: character_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionContextRequest"
      responses:
        "200":
          description: Persona snapshot augmented with immutable action facts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonaSnapshot"
        "404":
          description: No character exists with the requested identifier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: The request, action facts, or stored character are invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Persona or memory persistence is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/memories:
    get:
      operationId: listSophiaMemories
      summary: Inspect SophIA's selected memories
      responses:
        "200":
          description: Active memories and pending inferred candidates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryListResponse"
        "503":
          description: Continuous memory is disabled or unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: rememberSophiaMemory
      summary: Store an eligible selected memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RememberMemoryRequest"
      responses:
        "201":
          description: Memory created or matching memory refreshed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecordResponse"
        "409":
          description: Active-memory capacity has been reached.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Memory is invalid or ineligible.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Continuous memory is disabled or unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/memories/retrieve:
    post:
      operationId: retrieveSophiaMemories
      summary: Retrieve bounded memories relevant to a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrieveMemoryRequest"
      responses:
        "200":
          description: Deduplicated and sanitized relevant memories.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryListResponse"
        "422":
          description: Retrieval query is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Continuous memory is disabled or unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/memories/{memory_id}:
    patch:
      operationId: correctSophiaMemory
      summary: Correct a memory with optimistic concurrency
      parameters:
        - $ref: "#/components/parameters/MemoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CorrectMemoryRequest"
      responses:
        "200":
          description: Corrected memory; superseded content was removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecordResponse"
        "404":
          $ref: "#/components/responses/MemoryNotFound"
        "409":
          $ref: "#/components/responses/MemoryConflict"
        "422":
          $ref: "#/components/responses/MemoryInvalid"
        "503":
          $ref: "#/components/responses/MemoryUnavailable"
    delete:
      operationId: forgetSophiaMemory
      summary: Physically forget one memory
      parameters:
        - $ref: "#/components/parameters/MemoryId"
        - name: expected_revision
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "204":
          description: Memory content and representations were removed.
        "404":
          $ref: "#/components/responses/MemoryNotFound"
        "409":
          $ref: "#/components/responses/MemoryConflict"
        "422":
          $ref: "#/components/responses/MemoryInvalid"
        "503":
          $ref: "#/components/responses/MemoryUnavailable"
  /v1/memories/{memory_id}/confirm:
    post:
      operationId: confirmSophiaMemory
      summary: Confirm an inferred memory candidate
      parameters:
        - $ref: "#/components/parameters/MemoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionRequest"
      responses:
        "200":
          description: Candidate promoted to active memory.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecordResponse"
        "404":
          $ref: "#/components/responses/MemoryNotFound"
        "409":
          $ref: "#/components/responses/MemoryConflict"
        "422":
          $ref: "#/components/responses/MemoryInvalid"
        "503":
          $ref: "#/components/responses/MemoryUnavailable"
  /v1/memories/{memory_id}/complete:
    post:
      operationId: completeSophiaCommitment
      summary: Complete a commitment and start its retention window
      parameters:
        - $ref: "#/components/parameters/MemoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionRequest"
      responses:
        "200":
          description: Commitment marked complete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecordResponse"
        "404":
          $ref: "#/components/responses/MemoryNotFound"
        "409":
          $ref: "#/components/responses/MemoryConflict"
        "422":
          $ref: "#/components/responses/MemoryInvalid"
        "503":
          $ref: "#/components/responses/MemoryUnavailable"
components:
  parameters:
    MemoryId:
      name: memory_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 36
  responses:
    MemoryNotFound:
      description: Memory was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MemoryConflict:
      description: Memory revision is stale or capacity was reached.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MemoryInvalid:
      description: Memory operation is invalid or ineligible.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MemoryUnavailable:
      description: Continuous memory is disabled or unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthComponents:
      type: object
      additionalProperties: false
      required:
        - persona_engine
        - database
      properties:
        persona_engine:
          type: string
          const: ready
        database:
          type: string
          const: ready
    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - service
        - api_version
        - persona_schema_version
        - components
      properties:
        status:
          type: string
          const: ready
        service:
          type: string
          const: sophia-persona-engine
        api_version:
          type: string
          const: v1
        persona_schema_version:
          type: integer
          const: 1
        components:
          $ref: "#/components/schemas/HealthComponents"
    PersonaSnapshot:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - character_name
        - system_prompt
        - mood
        - dynamic
        - revision
      properties:
        schema_version:
          type: integer
          const: 1
        character_name:
          type: string
          minLength: 1
        system_prompt:
          type: string
          minLength: 1
        mood:
          type: string
          minLength: 1
        dynamic:
          type: boolean
        revision:
          type: integer
          minimum: 0
    TurnContextRequest:
      type: object
      additionalProperties: false
      required:
        - user_prompt
      properties:
        user_prompt:
          type: string
          minLength: 1
          maxLength: 8192
    ActionIntent:
      type: string
      enum:
        - open_url
        - remember_memory
      description: Allowlisted agent actions that Persona may describe but never execute.
    ActionStatus:
      type: string
      enum:
        - success
        - failed
        - denied
        - canceled
      description: Immutable outcome supplied by the authoritative agent core.
    ActionFactRequest:
      type: object
      additionalProperties: false
      required:
        - intent
        - arguments
        - status
        - detail
      properties:
        intent:
          $ref: "#/components/schemas/ActionIntent"
        arguments:
          type: object
          minProperties: 1
          maxProperties: 8
          additionalProperties:
            type: string
        status:
          $ref: "#/components/schemas/ActionStatus"
        detail:
          type: string
          minLength: 1
          maxLength: 512
    ActionContextRequest:
      type: object
      additionalProperties: false
      required:
        - user_prompt
        - action
      properties:
        user_prompt:
          type: string
          minLength: 1
          maxLength: 8192
        action:
          $ref: "#/components/schemas/ActionFactRequest"
    MemoryCategory:
      type: string
      enum:
        - preference
        - personal_fact
        - commitment
    MemoryOrigin:
      type: string
      enum:
        - explicit
        - inferred
    MemoryStatus:
      type: string
      enum:
        - active
        - candidate
    RememberMemoryRequest:
      type: object
      additionalProperties: false
      required:
        - category
        - content
        - origin
        - source_type
      properties:
        category:
          $ref: "#/components/schemas/MemoryCategory"
        content:
          type: string
          minLength: 1
        origin:
          $ref: "#/components/schemas/MemoryOrigin"
        source_type:
          type: string
          minLength: 1
          maxLength: 32
        source_timestamp:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        sensitive:
          type: boolean
          default: false
        purpose:
          anyOf:
            - type: string
              maxLength: 160
            - type: "null"
        confirmed:
          type: boolean
          default: false
    MemoryRecordResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - category
        - content
        - origin
        - status
        - source_type
        - source_timestamp
        - sensitive
        - revision
        - created_at
        - updated_at
      properties:
        id:
          type: string
        category:
          $ref: "#/components/schemas/MemoryCategory"
        content:
          type: string
        origin:
          $ref: "#/components/schemas/MemoryOrigin"
        status:
          $ref: "#/components/schemas/MemoryStatus"
        source_type:
          type: string
        source_timestamp:
          type: string
          format: date-time
        sensitive:
          type: boolean
        purpose:
          anyOf:
            - type: string
            - type: "null"
        revision:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        completed_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
    MemoryListResponse:
      type: object
      additionalProperties: false
      required:
        - memories
      properties:
        memories:
          type: array
          items:
            $ref: "#/components/schemas/MemoryRecordResponse"
    RetrieveMemoryRequest:
      type: object
      additionalProperties: false
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 8192
    CorrectMemoryRequest:
      type: object
      additionalProperties: false
      required:
        - content
        - expected_revision
      properties:
        content:
          type: string
          minLength: 1
        expected_revision:
          type: integer
          minimum: 1
    RevisionRequest:
      type: object
      additionalProperties: false
      required:
        - expected_revision
      properties:
        expected_revision:
          type: integer
          minimum: 1
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - detail
      properties:
        detail:
          $ref: "#/components/schemas/ErrorDetail"